Foi realizado a correção do código do arquivo FXML, foi feita alterações e correções visuais na tabela, craido um breve metodo que ao clicar no Button Carregar Documento, o arquivo CSV é apresentado no terminal do usuário, isso foi feito para já se ter uma base com a conexão com o Banco de Dados que virá em breve. Feita a alteração para que quando o cursor de mouse passase por cima dos botões, mudaria o cursor para pointer.
