+++ Cyber Nexus Pacer/Cyber Nexus - Pacer/src/main/resources/com/cyber/cybernexuspacer/fxml/cadastroTurma-view.fxml
@@ -1,8 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import javafx.scene.Cursor?>
 <?import javafx.scene.control.Button?>
 <?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ScrollBar?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
 <?import javafx.scene.image.Image?>
 <?import javafx.scene.image.ImageView?>
 <?import javafx.scene.layout.AnchorPane?>
@@ -11,51 +13,64 @@
 
 <AnchorPane maxHeight="600.0" maxWidth="800.0" minHeight="600.0" minWidth="800.0" prefHeight="694.0" prefWidth="800.0" style="-fx-background-color: #d9d9d9;" xmlns="http://javafx.com/javafx/22" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.cyber.cybernexuspacer.controller.CadastroDeTurmaController">
    <children>
+      <Label fx:id="label_top" layoutX="-23.0" layoutY="-7.0" prefHeight="40.0" prefWidth="822.0" style="-fx-background-color: #86B6DD;" text="        PRÓTIPO PACER">
+         <font>
+            <Font name="Arial Black" size="11.0" />
+         </font>
+      </Label>
       <Label fx:id="label_olaProfessor" layoutX="18.0" layoutY="60.0" prefHeight="30.0" prefWidth="161.0" text="Olá, Professor!">
          <font>
             <Font name="Arial Black" size="18.0" />
          </font>
       </Label>
-      <Label fx:id="label_cadastrodeturma" alignment="CENTER" layoutX="237.0" layoutY="40.0" prefHeight="39.0" prefWidth="378.0" style="-fx-background-color: #86B6DD;" text="CADASTRO DE TURMA" textAlignment="CENTER" textFill="WHITE">
+      <Label fx:id="label_cadastrodeturma" alignment="CENTER" layoutX="193.0" layoutY="52.0" prefHeight="39.0" prefWidth="456.0" style="-fx-background-color: #86B6DD;" text="CADASTRO DE TURMA" textAlignment="CENTER" textFill="WHITE">
          <font>
-            <Font name="Arial Bold" size="20.0" />
+            <Font name="Arial Black" size="30.0" />
          </font>
       </Label>
       <Button fx:id="button_sair" layoutX="25.0" layoutY="536.0" mnemonicParsing="false" prefHeight="30.0" prefWidth="98.0" style="-fx-background-color: #86B6DD;" text="SAIR" textFill="WHITE">
          <font>
             <Font name="Arial Black" size="12.0" />
          </font>
+         <cursor>
+            <Cursor fx:constant="HAND" />
+         </cursor>
       </Button>
-      <Button fx:id="button_confirmarAlunos" layoutX="614.0" layoutY="533.0" mnemonicParsing="false" onAction="#onClickbtnConfirmarAlunos" prefHeight="34.0" prefWidth="173.0" style="-fx-background-color: #369C06;" text="CONFIRMAR ALUNOS" textFill="WHITE">
+      <Button fx:id="button_confirmarAlunos" layoutX="614.0" layoutY="533.0" mnemonicParsing="false" onAction="#onClickConfirmarAlunos" prefHeight="34.0" prefWidth="173.0" style="-fx-background-color: #369C06;" text="CONFIRMAR ALUNOS" textFill="WHITE">
          <font>
             <Font name="Arial Black" size="12.0" />
          </font>
+         <cursor>
+            <Cursor fx:constant="HAND" />
+         </cursor>
       </Button>
-      <Button fx:id="button_carregarDocumento" layoutX="410.0" layoutY="533.0" mnemonicParsing="false" prefHeight="34.0" prefWidth="184.0" style="-fx-background-color: #FFB617;" text="CARREGAR DOCUMENTO" textFill="WHITE">
+      <Button fx:id="button_carregarDocumento" layoutX="410.0" layoutY="533.0" mnemonicParsing="false" onAction="#carregarPlanilha" prefHeight="34.0" prefWidth="184.0" style="-fx-background-color: #FFB617;" text="CARREGAR DOCUMENTO" textFill="WHITE">
          <font>
             <Font name="Arial Black" size="12.0" />
          </font>
+         <cursor>
+            <Cursor fx:constant="HAND" />
+         </cursor>
       </Button>
-      <Pane layoutX="13.0" layoutY="109.0" prefHeight="381.0" prefWidth="774.0" style="-fx-background-color: #FFFFFF;">
+      <ImageView fx:id="logo_fatec" fitHeight="100.0" fitWidth="105.0" layoutX="688.0" layoutY="24.0" pickOnBounds="true" preserveRatio="true">
+         <image>
+            <Image url="@../../../../../../../../../../../../api/src/main/resources/images/ada.png" />
+         </image>
+      </ImageView>
+      <AnchorPane layoutX="2.0" layoutY="104.0" prefHeight="402.0" prefWidth="791.0">
          <children>
-            <Pane fx:id="pane_planilha" layoutX="17.0" layoutY="9.0" prefHeight="355.0" prefWidth="735.0" style="-fx-background-color: #d9d9d9d9;" />
-            <ScrollBar fx:id="scrollBar" layoutX="751.0" layoutY="10.0" orientation="VERTICAL" prefHeight="355.0" prefWidth="19.0" />
-               <ImageView fitHeight="323.0" fitWidth="911.0" layoutX="31.0" layoutY="23.0" pickOnBounds="true" preserveRatio="true">
-                  <image>
-                     <Image url="@../imagens/PLANE.jpg" />
-                  </image>
-               </ImageView>
+            <Pane layoutX="6.0" layoutY="7.0" prefHeight="372.0" prefWidth="778.0" style="-fx-background-color: #FFFFFF;">
+               <children>
+                  <TableView fx:id="tabela" accessibleText="TABLE_VIEW" cache="true" editable="true" prefHeight="393.0" prefWidth="778.0">
+                    <columns>
+                      <TableColumn fx:id="coluna_Aluno" prefWidth="247.999981880188" text="ALUNO" />
+                        <TableColumn fx:id="coluna_Email" prefWidth="265.5999755859375" text="EMAIL" />
+                        <TableColumn fx:id="coluna_Grupo" minWidth="0.0" prefWidth="261.60009765625" text="GRUPO" />
+                    </columns>
+                  </TableView>
+               </children>
+            </Pane>
          </children>
-      </Pane>
-      <ImageView fitHeight="60.0" fitWidth="107.0" layoutX="668.0" layoutY="22.0" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@../imagens/fatec.png" />
-         </image>
-      </ImageView>
-      <ImageView blendMode="MULTIPLY" fitHeight="60.0" fitWidth="150.0" layoutX="8.0" layoutY="1.0" opacity="0.4" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@../imagens/logo Cyber Nexus.png" />
-         </image>
-      </ImageView>
+      </AnchorPane>
    </children>
 </AnchorPane>
